Extracting Cosmetic Ingredients & Analysis/requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=9.5.0
numpy>=1.23.0
regex>=2023.6.3
sqlalchemy
psycopg2-binary
